refactor(UserPage): remove props drilling tech debt
